refactor: split tasks into Typer CLI
