+++ requirements.txt
@@ -16,6 +16,7 @@ charset-normalizer==3.4.2
 click==8.2.1
 cloudpickle==3.1.1
 colorlog==6.9.0
+typer==0.12.3
 datasets==4.0.0
 decorator==5.2.1
 dill==0.3.8
