+++ tests/test_parse_codex_tasks.py
@@ -5,8 +5,8 @@ from auto.html_utils import parse_codex_tasks
 def test_parse_codex_tasks():
     html = pathlib.Path("tests/fixtures/dom.html").read_text()
     tasks = parse_codex_tasks(html)
-    assert len(tasks) == 6
+    assert len(tasks) == 8
     statuses = [t["status"] for t in tasks]
-    assert statuses.count("Merged") == 2
+    assert statuses.count("Merged") == 3
     assert statuses.count("Open") == 1
-    assert sum(1 for t in tasks if not t["code"]) == 3
+    assert sum(1 for t in tasks if not t["code"]) == 4
